desktop-monitoring/requirements.txt
matplotlib>=3.8.0
mss>=9.0.1
openpyxl>=3.1.5
Pillow>=10.4.0
psutil>=6.0.0
pyperclip>=1.9.0
pywin32>=306
screeninfo>=0.8.1
opencv-python>=4.10.0
python-dotenv>=1.0.1
